+++ telegram/choseninlineresult.py
@@ -66,18 +66,18 @@ class ChosenInlineResult(TelegramObject):
 
         self._id_attrs = (self.result_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        data = super(ChosenInlineResult, ChosenInlineResult).de_json(data, bot)
+        data = super(ChosenInlineResult, cls).de_json(data, bot)
         # Required
         data['from_user'] = User.de_json(data.pop('from'), bot)
         # Optionals
         data['location'] = Location.de_json(data.get('location'), bot)
 
-        return ChosenInlineResult(**data)
+        return cls(**data)
 
     def to_dict(self):
         data = super(ChosenInlineResult, self).to_dict()