+++ telegram/files/audio.py
@@ -64,9 +64,9 @@ class Audio(TelegramObject):
 
         self._id_attrs = (self.file_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        return Audio(**data)
+        return cls(**data)