+++ telegram/files/file.py
@@ -58,12 +58,12 @@ class File(TelegramObject):
 
         self._id_attrs = (self.file_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        return File(bot=bot, **data)
+        return cls(bot=bot, **data)
 
     def download(self, custom_path=None, out=None, timeout=None):
         """