+++ telegram/files/location.py
@@ -42,9 +42,9 @@ class Location(TelegramObject):
 
         self._id_attrs = (self.longitude, self.latitude)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        return Location(**data)
+        return cls(**data)