+++ telegram/files/photosize.py
@@ -49,20 +49,20 @@ class PhotoSize(TelegramObject):
 
         self._id_attrs = (self.file_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        return PhotoSize(**data)
+        return cls(**data)
 
-    @staticmethod
-    def de_list(data, bot):
+    @classmethod
+    def de_list(cls, data, bot):
         if not data:
             return []
 
         photos = list()
         for photo in data:
-            photos.append(PhotoSize.de_json(photo, bot))
+            photos.append(cls.de_json(photo, bot))
 
         return photos