+++ telegram/files/video.py
@@ -66,13 +66,13 @@ class Video(TelegramObject):
 
         self._id_attrs = (self.file_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        data = super(Video, Video).de_json(data, bot)
+        data = super(Video, cls).de_json(data, bot)
 
         data['thumb'] = PhotoSize.de_json(data.get('thumb'), bot)
 
-        return Video(**data)
+        return cls(**data)