+++ telegram/files/videonote.py
@@ -52,13 +52,13 @@ class VideoNote(TelegramObject):
 
         self._id_attrs = (self.file_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        data = super(VideoNote, VideoNote).de_json(data, bot)
+        data = super(VideoNote, cls).de_json(data, bot)
 
         data['thumb'] = PhotoSize.de_json(data.get('thumb'), bot)
 
-        return VideoNote(**data)
+        return cls(**data)