+++ telegram/files/voice.py
@@ -49,11 +49,11 @@ class Voice(TelegramObject):
 
         self._id_attrs = (self.file_id,)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        data = super(Voice, Voice).de_json(data, bot)
+        data = super(Voice, cls).de_json(data, bot)
 
-        return Voice(**data)
+        return cls(**data)