+++ telegram/replykeyboardremove.py
@@ -18,7 +18,6 @@
 # along with this program.  If not, see [http://www.gnu.org/licenses/].
 """This module contains an object that represents a Telegram ReplyKeyboardRemove."""
 from telegram import ReplyMarkup
-from telegram.utils.deprecate import warn_deprecate_obj
 
 
 class ReplyKeyboardRemove(ReplyMarkup):
@@ -55,23 +54,9 @@ class ReplyKeyboardRemove(ReplyMarkup):
         # Optionals
         self.selective = bool(selective)
 
-    @staticmethod
-    def de_json(data, bot):
+    @classmethod
+    def de_json(cls, data, bot):
         if not data:
             return None
 
-        return ReplyKeyboardRemove(**data)
-
-
-class ReplyKeyboardHide(object):
-    """ReplyKeyboardHide
-
-    Note:
-        Deprecated.
-    """
-
-    def __new__(cls, hide_keyboard=True, selective=False, **kwargs):
-        warn_deprecate_obj(ReplyKeyboardHide.__name__, ReplyKeyboardRemove.__name__)
-        obj = ReplyKeyboardRemove.__new__(ReplyKeyboardRemove, selective)
-        obj.__init__(selective)
-        return obj
+        return cls(**data)