+++ CHANGES.rst
@@ -7,9 +7,12 @@ Changes
 - Improved filters for user_id/username/chat.
 - Internal restructure of files.
 - Improved unitests.
-- Fully support Bot API 3.1.
 - Fully support Bot API 3.2
 - Modified docstrings
+- Remove deprecated ``telegram.Emoji``.
+- Remove deprecated ``Botan`` import from ``utils``.
+- Remove deprecated ``ReplyKeyboardHide``.
+- Remove deprecated ``edit_message`` argument of `bot.set_game_score``.
 
 **2017-06-18**